% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_files}
\alias{download_files}
\title{Download files from file list to a destination folder}
\usage{
download_files(
  files_df,
  dest_folder,
  filename_prefix = NULL,
  skip_exists = FALSE
)
}
\arguments{
\item{files_df}{(data.frame) File list collected from \code{\link{get_file_list}}.}

\item{dest_folder}{(character) Destination local folder to save files.}

\item{filename_prefix}{(character) Prefix for file names.}

\item{skip_exists}{(boolean) Skips downloading if the file}
}
\value{
None.
}
\description{
Download files from file list collected from \code{\link{get_file_list}}
to a destination folder
}
\details{
The file links from  \code{\link{get_file_list}} are valid for 24 hours.
If download process passes the 24 hours period,
the \code{files_df} will be no longer valid
and \code{\link{get_file_list}} must be called again to collect new
\code{files_df}. Then run this function again with \code{skip_exists = TRUE} option.
}
\examples{
# not run
download_files(files_df, "C:/temp", "advan_wp_", skip_exists = FALSE)

}
\references{
None
}
\seealso{
\code{\link{download_files0}}
}
\author{
Dewey Data Inc.
}
