% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_files0}
\alias{download_files0}
\title{Download files with API key and product path to a destination folder}
\usage{
download_files0(
  apikey,
  product_path,
  dest_folder,
  filename_prefix = NULL,
  skip_exists = TRUE
)
}
\arguments{
\item{apikey}{(character) API Key.}

\item{product_path}{(character) API endpoint or Product ID.}

\item{dest_folder}{(character) Destination local folder to save files.}

\item{filename_prefix}{(character) Prefix for file names.}

\item{skip_exists}{(boolean) Skips downloading if the file}
}
\value{
None.
}
\description{
Download files with API key and product path to a destination folder.
}
\details{
The file links inside this function are valid for 24 hours.
If download process passes the 24 hours period,
this function must be called again with \code{skip_exists = TRUE} option.
}
\examples{
# not run
download_files0(apikey_, product_path, "C:/temp")

}
\references{
None
}
\seealso{
\code{\link{download_files}}
}
\author{
Dewey Data Inc.
}
