% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_files1}
\alias{download_files1}
\title{Download files with API key and product path to a destination folder}
\usage{
download_files1(
  apikey,
  product_path,
  dest_folder,
  start_date = NULL,
  end_date = NULL,
  filename_prefix = NULL,
  skip_exists = FALSE
)
}
\arguments{
\item{apikey}{(character) API Key.}

\item{product_path}{(character) API endpoint or Product ID.}

\item{dest_folder}{(character) Destination local folder to save files.}

\item{start_date}{(character) Data start date character for files in the form of "2021-07-01".
Default is NULL ("1000-01-01"), which indicates no limit.}

\item{end_date}{(character) Data end date character for files in the form of "2023-08-21".
Default is NULL ("9999-12-31"), which indicates no limit.}

\item{filename_prefix}{(character) Prefix for file names.}

\item{skip_exists}{(boolean) Skips downloading if the file exists. Default is FALSE.}
}
\value{
None.
}
\description{
Download files with API key and product path to a destination folder.
}
\details{
None.
}
\examples{
# not run
download_files1(apikey_, product_path, "C:/temp")

}
\references{
None
}
\seealso{
\code{\link{download_files}}, \code{\link{download_files0}}
}
\author{
Dewey Data Inc.
}
