% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_file_list}
\alias{get_file_list}
\title{Collects file list from server}
\usage{
get_file_list(
  apikey,
  product_path,
  start_page = 1,
  end_page = Inf,
  partition_key_after = NA,
  partition_key_before = NA,
  print_info = T
)
}
\arguments{
\item{apikey}{(character) API Key.}

\item{product_path}{(character) API endpoint or Product ID.}

\item{start_page}{(integer) Start page of file list. Default is 1.}

\item{end_page}{(integer) End page of file list. Default is Inf.}

\item{partition_key_after}{(character) Start date character for files in the form of "20021-07-01".
Default is NA ("1000-01-01"), which indicates no limit.}

\item{partition_key_before}{(character) End date character for files in the form of "2023-08-21".
Default is NA ("9999-12-31"), which indicates no limit.}

\item{print_info}{(logical) Print file list information. Default is TRUE.}
}
\value{
A data.frame object contains files information.
  \item{index}{file index starts from 1.}
  \item{page}{page of the file list.}
  \item{link}{download link.}
  \item{partition_key}{date field to partiton files.}
  \item{file_name}{file name.}
  \item{file_extension}{file extension (e.g. .csv.gz.)}
  \item{download_lik}{same as \code{link}.}
}
\description{
Collects the file list information from data server.
}
\details{
None
}
\examples{
# not run
files_df = get_file_list(apikey_, product_path_,
           start_page = 1, end_page = 3, print_info = T)

}
\references{
None
}
\seealso{
\code{\link{slice_files_df}}, \code{\link{read_sample_data}}
}
\author{
Dewey Data Inc.
}
