% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_meta}
\alias{get_meta}
\title{Get metadata from server}
\usage{
get_meta(apikey, product_path, print_meta = TRUE)
}
\arguments{
\item{apikey}{(character) API Key.}

\item{product_path}{(character) API endpoint or Product ID.}

\item{print_meta}{(logical) Print metadata. Default is TRUE.}
}
\value{
A data.frame object contains metadata.
 \item{total_files}{total number of files.}
 \item{total_size_MB}{total size of files in MB.}
 \item{partition_column}{data partition column name.}
 \item{partition_aggregation}{aggregation frequency.}
 \item{min_partition_key}{min date available.}
 \item{max_partition_key}{max date available.}
}
\description{
Get metadata from server
}
\details{
None
}
\examples{
# not run
meta_df = get_meta(apikey_, product_path_, print_meta = TRUE)

}
\references{
None
}
\author{
Dewey Data Inc.
}
