% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyval_populate.R
\name{keyval_to_df}
\alias{keyval_to_df}
\title{Convert JSON data rows to data.frame}
\usage{
keyval_to_df(keyval, keys = NULL, type = "json", asis = F)
}
\arguments{
\item{keyval}{(character) key-value pairs text (JSON, etc.).}

\item{keys}{(character) keys (vector) to populate.}

\item{type}{(character) c("json", "pass_by").}

\item{asis}{(boolean) Populate all keys in the data without error checking.}
}
\value{
A data.frame object contains key-values.
}
\description{
Convert JSON data rows to data.frame.
}
\details{
None
}
\examples{
# not run
aa = keyval_to_df(sample_data$POPULARITY_BY_DAY,
                  c("Friday", "Monday", "Whatday"), asis = F)

}
\references{
None
}
\author{
Dewey Data Inc.
}
