% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{read_local_data}
\alias{read_local_data}
\title{Read local data into memory from a path}
\usage{
read_local_data(path, nrows = -1)
}
\arguments{
\item{nrows}{(integer) Number of rows to read. Default is -1 (all).}

\item{apikey}{(character) Path to a .csv.gz file.}
}
\value{
A data.frame object contains data.
}
\description{
Read local data into memory from a path
}
\details{
This can be slow. Recommend to use fread function at data.table package.
}
\examples{
# not run
sample_data = read_sample_data(files_df$link[1], nrows = 100)

}
\references{
None
}
\seealso{
None
}
\author{
Dewey Data Inc.
}
