% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{read_sample0}
\alias{read_sample0}
\title{Read sample data into memory from a URL}
\usage{
read_sample0(apikey, product_path, nrows = 100)
}
\arguments{
\item{apikey}{(character) API key.}

\item{product_path}{(character) API endpoint.}

\item{nrows}{(integer) Number of rows to read. Default is 100.}
}
\value{
A data.frame object contains data.
}
\description{
Read first file data into memory using API key and product path.
}
\details{
None
}
\examples{
# not run
sample_data0 = read_sample0(apikey_, pp_advan_wp, 100)

}
\references{
None
}
\seealso{
\code{\link{get_file_list}}, \code{\link{read_sample_data}}
}
\author{
Dewey Data Inc.
}
