% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{slice_files_df}
\alias{slice_files_df}
\title{Slice files_df from \code{get_file_list}}
\usage{
slice_files_df(files_df, start_date, end_date = NULL)
}
\arguments{
\item{files_df}{(data.frame) \code{files_df} from \code{get_file_list}.}

\item{start_date}{(character) Start date character for files in the form of "2023-08-21".}

\item{end_date}{(character) End date character for files in the form of "2023-08-21".
Default is \code{NULL}, which indicates slice to the last date.}
}
\value{
None.
}
\description{
Slice \code{files_df} from \code{get_file_list} for specific data range
of from start_date to end_date.
For example, start_date = "2023-08-14", end_date = "2023-08-21".
}
\details{
None
}
\examples{
# not run
sliced_files_df = slice_files_df(files_df, "2023-09-01", "2023-09-10")

}
\references{
None
}
\seealso{
\code{\link{get_file_list}}
}
\author{
Dewey Data Inc.
}
