% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xlview}
\alias{xlview}
\title{View data in Excel}
\usage{
xlview(data)
}
\arguments{
\item{data}{data.frame.}
}
\value{
None.
}
\description{
View data in Excel
}
\details{
!! WARNING!!
This will write the data to a temp file and the file will not be removed
and it is vulnerable for security threats.
Thus, do not use this function for any proprietary data.
Also, only one Excel view can be opened at one time.
}
\examples{
# not run
xlview(data.frame(a = 3, b = 4))

}
\references{
None
}
\seealso{
None
}
\author{
Dewey Data Inc.
}
