% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deweydatar_package.R
\docType{package}
\name{deweydatar-package}
\alias{deweydatar-package}
\title{deweydatar: An R package for Dewey Data Inc.}
\description{
This package includes helper functions for Dewey Data.
}
\note{
This is a beta version.
}
\references{
None
}
\author{
Dewey Data Inc.
}
